Fix recycle dependency eslint and fix console error
